refactor: split InfiniteScroll class - refactor types - add handlers, utils
