+++ src/InfiniteScroll/index.ts
@@ -1,0 +1,4 @@
+export * from './handlers';
+export * from './types';
+export * from './utils';
+export { InfiniteScroll } from './InfiniteScroll';
