+++ src/InfiniteScroll/types/enums.ts
@@ -1,0 +1,11 @@
+export enum ScrollAxisName {
+  VERTICAL = 'vertical',
+  HORIZONTAL = 'horizontal',
+}
+
+export enum ScrollDirection {
+  UP = 'up',
+  DOWN = 'down',
+  LEFT = 'left',
+  RIGHT = 'right',
+}
