+++ src/InfiniteScroll/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './getValidScrollValue';
+export * from './getMultiplierSign';
