+++ src/index.ts
@@ -1,7 +1,2 @@
-import useInfiniteScroll from './useInfiniteScroll';
-
-export { default as InfiniteScroll } from './InfiniteScroll';
-export { default as useInfiniteScroll } from './useInfiniteScroll';
-export * from './types';
-
-export default useInfiniteScroll;
+export * from './InfiniteScroll';
+export * from './useInfiniteScroll';
